refactor: global.util -> 각각의 util 패키지로 이동
