+++ src/main/java/com/hanium/mom4u/global/crawling/listener/CrawlEventListener.java
@@ -1,6 +1,6 @@
 package com.hanium.mom4u.global.crawling.listener;
 
-import com.hanium.mom4u.global.util.SeleniumCrawler;
+import com.hanium.mom4u.global.crawling.service.SeleniumCrawler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
