+++ src/main/java/com/hanium/mom4u/global/security/service/KakaoLoginService.java
@@ -9,8 +9,8 @@ import com.hanium.mom4u.global.security.dto.response.KakaoTokenDto;
 import com.hanium.mom4u.global.security.dto.response.KakaoUserInfoResponseDto;
 import com.hanium.mom4u.global.security.dto.response.LoginResponseDto;
 import com.hanium.mom4u.global.security.jwt.JwtTokenProvider;
-import com.hanium.mom4u.global.util.KakaoUtil;
-import com.hanium.mom4u.global.util.RefreshTokenUtil;
+import com.hanium.mom4u.global.security.util.KakaoUtil;
+import com.hanium.mom4u.global.security.util.RefreshTokenUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
