+++ src/main/java/com/hanium/mom4u/global/security/service/NaverLoginService.java
@@ -8,8 +8,8 @@ import com.hanium.mom4u.domain.member.repository.MemberRepository;
 import com.hanium.mom4u.global.security.dto.response.LoginResponseDto;
 import com.hanium.mom4u.global.security.dto.response.NaverProfileResponseDto;
 import com.hanium.mom4u.global.security.jwt.JwtTokenProvider;
-import com.hanium.mom4u.global.util.NaverUtil;
-import com.hanium.mom4u.global.util.RefreshTokenUtil;
+import com.hanium.mom4u.global.security.util.NaverUtil;
+import com.hanium.mom4u.global.security.util.RefreshTokenUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
