+++ src/main/java/com/hanium/mom4u/global/security/util/KakaoUtil.java
@@ -1,4 +1,4 @@
-package com.hanium.mom4u.global.util;
+package com.hanium.mom4u.global.security.util;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
